feat: chunked log segment manager
